+++ tests/test_is_protocol.py
@@ -1,3 +1,6 @@
+# SPDX-FileCopyrightText: 2021-2023 Constantine Evans <qslib@mb.costi.net>
+# SPDX-License-Identifier: EUPL-1.2
+
 import asyncio
 import logging