+++ tests/test_protocol.py
@@ -1,5 +1,5 @@
-# SPDX-FileCopyrightText: 2021-2022 Constantine Evans <const@costi.eu>
-# SPDX-License-Identifier: AGPL-3.0-only
+# SPDX-FileCopyrightText: 2021-2023 Constantine Evans <qslib@mb.costi.net>
+# SPDX-License-Identifier: EUPL-1.2
 
 import pathlib